feat(tilemap): :bug: Fix events with tiles
