+++ src/components/tilemap/tiles/atlas.rs
@@ -15,7 +15,12 @@
 // You should have received a copy of the GNU General Public License
 // along with Luminol.  If not, see <http://www.gnu.org/licenses/>.
 
-use super::{AUTOTILE_HEIGHT, MAX_SIZE, TILESET_WIDTH, TOTAL_AUTOTILE_HEIGHT, UNDER_HEIGHT};
+use super::autotile_ids::AUTOTILES;
+use super::Quad;
+use super::{
+    AUTOTILE_AMOUNT, AUTOTILE_HEIGHT, MAX_SIZE, TILESET_WIDTH, TOTAL_AUTOTILE_HEIGHT, UNDER_HEIGHT,
+};
+
 use crate::prelude::*;
 
 use image::GenericImageView;
@@ -208,6 +213,95 @@ impl Atlas {
             tileset_height: tileset_img.height(),
             autotile_frames,
         })
+    }
+
+    pub fn calc_quads(&self, tile: i16, x: usize, y: usize, z: usize, quads: &mut Vec<Quad>) {
+        // There are 4 cases we need to handle here:
+        match tile {
+            // The tile is blank
+            0..=47 => {}
+            // The tile is an autotile
+            48..=384 => {
+                let autotile_id = (tile / 48) - 1;
+                for s_a in 0..2 {
+                    for s_b in 0..2 {
+                        let pos = egui::Rect::from_min_size(
+                            egui::pos2(
+                                x as f32 * 32. + (s_a as f32 * 16.),
+                                y as f32 * 32. + (s_b as f32 * 16.),
+                            ),
+                            egui::vec2(16., 16.),
+                        );
+
+                        let ti = AUTOTILES[tile as usize % 48][s_a + (s_b * 2)];
+                        // let tile_x = ti % 6;
+                        let tile_x = (ti % 6 * 16) as f32;
+
+                        let tile_y =
+                            (ti / 6 * 16) as f32 + (AUTOTILE_HEIGHT * autotile_id as u32) as f32;
+
+                        let tex_coords = egui::Rect::from_min_size(
+                            egui::pos2(tile_x, tile_y),
+                            egui::vec2(16., 16.),
+                        );
+
+                        quads.push(Quad::new(pos, tex_coords, z as f32));
+                    }
+                }
+            }
+            // The tileset does not wrap
+            tile if self.tileset_height + TOTAL_AUTOTILE_HEIGHT <= MAX_SIZE => {
+                let tile = tile - 384;
+
+                let pos = egui::Rect::from_min_size(
+                    egui::pos2(x as f32 * 32., y as f32 * 32.),
+                    egui::vec2(32., 32.),
+                );
+
+                let tile_x = (tile % 8) as f32 * 32.;
+                let tile_y = (tile as u32 / 8 + AUTOTILE_AMOUNT * 4) as f32 * 32.;
+                let tex_coords =
+                    egui::Rect::from_min_size(egui::pos2(tile_x, tile_y), egui::vec2(32., 32.));
+
+                quads.push(Quad::new(pos, tex_coords, z as f32));
+            }
+            // The tileset *does* wrap
+            tile => {
+                let tile = tile - 384;
+
+                let pos = egui::Rect::from_min_size(
+                    egui::pos2(x as f32 * 32., y as f32 * 32.),
+                    egui::vec2(32., 32.),
+                );
+
+                let tile_x = (tile as u32 % 8) * 32;
+                let tile_y = (tile as u32 / 8 + AUTOTILE_AMOUNT * 4) * 32;
+
+                let h1 = AUTOTILE_HEIGHT + UNDER_HEIGHT * self.columns_under;
+                let mut wrapped_x;
+                let mut wrapped_y;
+                if tile_y < h1 {
+                    // you're underneath the autotiles
+                    wrapped_x = tile_x + (tile_y - AUTOTILE_HEIGHT) / UNDER_HEIGHT * TILESET_WIDTH;
+                    wrapped_y = AUTOTILE_HEIGHT + (tile_y - AUTOTILE_HEIGHT) % UNDER_HEIGHT;
+                } else {
+                    // you're to the right
+                    // first wrap immediately to the right of the autotiles
+                    wrapped_x = tile_x + self.autotile_width;
+                    wrapped_y = tile_y - h1;
+                    // then wrap further
+                    wrapped_x += wrapped_y / MAX_SIZE * TILESET_WIDTH;
+                    wrapped_y %= MAX_SIZE;
+                }
+
+                let tex_coords = egui::Rect::from_min_size(
+                    egui::pos2(wrapped_x as f32, wrapped_y as f32),
+                    egui::vec2(32., 32.),
+                );
+
+                quads.push(Quad::new(pos, tex_coords, z as f32));
+            }
+        }
     }
 
     pub fn bind<'rpass>(&'rpass self, render_pass: &mut wgpu::RenderPass<'rpass>) {
